Moved commands around into sub-folders, helpers eliminated global.py
